pyproject.toml
[project]
name = "solana-token-transfer"
version = "0.1.0"
description = "Transfer USDC and SOL tokens on Solana blockchain"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "solana>=0.34.0",
    "solders>=0.21.0",
    "python-dotenv>=1.0.0",
    "anchorpy>=0.19.0",
    "httpx[socks]>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = []
